Add lineEnding property to code formatter setting
This is because the code formatter plugin automatically choose lineEnding characters according to the operation system. On windows, it will be CRLF, which is different from the linux/macos. If we do a -Dformat on windows, it will mark every lines of all the text files as modified.
By forcing the lineEnding property to LF, it will convert all CRLF to LF during the formatting step and save it to the files. This behavior is consistent with the suggested core.autocrlf setting for projects with developpers accross different platforms.