Improve Eclipse Java formatter, stop wrapping lines
The Eclipse formatter currently wraps lines automatically which makes the code harder to read in common use-cases like the use of builders.

Ex:

```java
Record.builder()
  .withType(RECORD)
  .withEntry(entry1)
  .withEntry(entry2)
  .build()
```

Becomes:

```java
Record.builder().withType(RECORD).withEntry(entry1).withEntry(entry2).build()
```

This is worked around in a lot of places by adding an empty comment at the end of lines that should not be wrapped, which shows that it's time for the rule to evolve.

This changes makes the formatter preserve wrapping, one is free to remove them manually where necessary.